pyproject.toml
[project]
name = "email-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.6",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "pywin32>=310",
]
